requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.3
django-jazzmin==3.0.1
django-ninja==1.3.0
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
python-dotenv==1.0.1
sqlparse==0.5.2
typing_extensions==4.12.2
